+++ src/visuals/renderer.cpp
@@ -49,17 +49,17 @@ void Renderer::RenderTexture(SDL_Texture* texture, int x, int y, int width,
 void Renderer::RenderTexture(SDL_Texture* texture,
                              float x,
                              float y,
-                             int width,
-                             int height) noexcept {
+                             float width,
+                             float height) noexcept {
   if ((texture != nullptr) && (width > 0) && (height > 0)) {
-    SDL_FRect dst = {x, y, static_cast<float>(width), static_cast<float>(height)};
+    SDL_FRect dst = {x, y, width, height};
     SDL_RenderCopyF(renderer, texture, nullptr, &dst);
   }
 }
 
-void Renderer::RenderFillRect(float x, float y, int width, int height) noexcept {
+void Renderer::RenderFillRect(float x, float y, float width, float height) noexcept {
   if ((width > 0) && (height > 0)) {
-    SDL_FRect rect = {x, y, static_cast<float>(width), static_cast<float>(height)};
+    SDL_FRect rect = {x, y, width, height};
     SDL_RenderFillRectF(renderer, &rect);
   }
 }
@@ -71,9 +71,9 @@ void Renderer::RenderFillRect(int x, int y, int width, int height) noexcept {
   }
 }
 
-void Renderer::RenderRect(float x, float y, int width, int height) noexcept {
+void Renderer::RenderRect(float x, float y, float width, float height) noexcept {
   if ((width > 0) && (height > 0)) {
-    SDL_FRect rect = {x, y, static_cast<float>(width), static_cast<float>(height)};
+    SDL_FRect rect = {x, y, width, height};
     SDL_RenderDrawRectF(renderer, &rect);
   }
 }
